remove duplicate method run_subtests
